Changes to the index file and css added
Many additions have been seen all the way through. Areas to be referred back to:
The margins are not precisely as they are in the mock up.
The attempts to change the checkboxes need cleaning up. Not precise colours and the checked version hangs too low.
Radio buttons are still on default settings.